restructure code to enable search bar to show above hero section and for watchlist components to be consistent with rest of the other components
