+++ Watchlist-Project/src/App.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import Homepage from './pages/Homepage'
@@ -8,7 +8,6 @@ import WatchList from './pages/WatchList';
 
 
 function App() {
-  const [count, setCount] = useState(0)
 
   return (
     <>
@@ -16,8 +15,8 @@ function App() {
 
       <Router>
         <Routes>
-          <Route path='/' element={<Homepage />}/> 
-          <Route path='MyReviews' element={<MyReviews />}/> 
+          <Route path='/' element={<Homepage/>}/> 
+          <Route path='MyReviews' element={<MyReviews/>}/> 
           <Route path='Watchlist' element={<WatchList />}/> 
           
       </Routes>
