+++ Watchlist-Project/src/components/WatchlistComponent.jsx
@@ -1,39 +1,24 @@
 // WatchListComponent.jsx
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import ReviewModalWithButton from './ReviewModalWithButton';
 import RemoveButton from './RemoveButton';
 
-const WatchList = () => {
-  const [watchlist, setWatchlist] = useState([]);
-
-  useEffect(() => {
-    const storedWatchlist = JSON.parse(localStorage.getItem('watchlist'));
-    if (storedWatchlist) {
-      setWatchlist(storedWatchlist);
-    }
-  }, []);
-
-  const removeFromWatchlist = (id) => {
-    const updatedWatchlist = watchlist.filter(item => item.id !== id);
-    setWatchlist(updatedWatchlist);
-    localStorage.setItem('watchlist', JSON.stringify(updatedWatchlist));
-  };
-
+const WatchList = (props) => {
   return (
     <div>
       <h2>Watchlist</h2>
-      <ul>
-        {watchlist.map(item => (
-          <li key={item.id}>
+      <div>
+        {props.watchlist.map(item => (
+          <div key={item.id}>
             <h3>{item.title}</h3>
             {item.poster && <img src={item.poster} alt={item.title} />}
             <div>
               <ReviewModalWithButton show={item} />
-              <RemoveButton showId={item.id} removeFromWatchlist={removeFromWatchlist} />
+              <RemoveButton showId={item.id} removeFromWatchlist={props.removeFromWatchlist} />
             </div>
-          </li>
+          </div>
         ))}
-      </ul>
+      </div>
     </div>
   );
 };
